refactor(temporal): infer SyncActivities from createActivities return type
Committed-By-Agent: claude
