+++ apps/service/src/__tests__/temporal-workflow.test.ts
@@ -2,7 +2,8 @@ import { describe, it, expect, beforeAll, afterAll } from 'vitest'
 import { TestWorkflowEnvironment } from '@temporalio/testing'
 import { Worker } from '@temporalio/worker'
 import path from 'node:path'
-import type { SyncActivities, RunResult } from '../temporal/types.js'
+import type { SyncActivities } from '../temporal/activities.js'
+import type { RunResult } from '../temporal/types.js'
 
 // workflowsPath must point to compiled JS (Temporal bundles it for V8 sandbox)
 const workflowsPath = path.resolve(process.cwd(), 'dist/temporal/workflows.js')
